+++ R/gips_class.R
@@ -2005,8 +2005,6 @@ BIC.gips <- function(object, ...){
 #'
 #' @param g An object of class "gips";
 #'     a result of a `find_MAP(return_probabilities = TRUE)`.
-#' @param sorted Logical; for `TRUE` (default) the output
-#'     will be sorted according to the probability.
 #'
 #' @returns Returns a numeric vector, calculated values of probabilities.
 #' Names contains permutations this probability represent.
@@ -2031,7 +2029,7 @@ BIC.gips <- function(object, ...){
 #' )
 #'
 #' get_probabilities_from_gips(g_map)
-get_probabilities_from_gips <- function(g, sorted = TRUE) {
+get_probabilities_from_gips <- function(g) {
   validate_gips(g)
 
   if (is.null(attr(g, "optimization_info"))) {
@@ -2053,13 +2051,7 @@ get_probabilities_from_gips <- function(g, sorted = TRUE) {
     ))
   }
 
-  out <- attr(g, "optimization_info")[["post_probabilities"]]
-  
-  if (sorted){
-    out <- sort(out, decreasing = TRUE)
-  }
-  
-  out
+  attr(g, "optimization_info")[["post_probabilities"]]
 }