+++ tests/testthat/test-gips_class.R
@@ -1287,9 +1287,6 @@ test_that("get_probabilities_from_gips works", {
   # sorted
   probs <- get_probabilities_from_gips(g_map)
   expect_equal(order(probs, decreasing = TRUE), c(1,2))
-  
-  probs <- get_probabilities_from_gips(g_map, sorted = FALSE)
-  expect_equal(order(probs, decreasing = TRUE), c(2,1))
 })
 
 test_that("forget_perms works properly", {